#4 feat: 챌린지 네트워크 모델 생성
